+++ CHANGES.rst
@@ -3,8 +3,10 @@ Change Log
 
 2.4.0
 -----
+
 * Pickup Service usingv11 WSDL (hornedbull)
-* TODO: Added examples, documentation and unit tests for Pickup Service.
+* Added documentation and unit tests for Pickup Service. (radzhome)
+* Update package data to include tools (noodlebreak)
 
 2.3.1
 -----